feat(record): support enabling VAAPI with CQP 5
Add support for VAAPI while encoding, using FFmpeg's h264_vaapi encoder.
To enable it, the --vaapi or --hw flags can be used. There are no real
plans to support other types of hardware acceleration, such as VDPAU and
NVENC, for the time being. At least not from my part.

A CQP of 5 was used to prevent the file size from reaching astronomical
measures, but retain a decent standard of quality. This might become
configurable in the future, or the value might change.