fix(fzfmenu): use exec instead of nohup in fzfmenu_run
This makes it so the process doesn't hang around even if /bin/sh is a
link to Bash. The nohup solution was something I came up with at some
point since exec wasn't doing the trick, but it seems the inverse holds
true for Bash.

Go figure. This could change at some point. Works for me™.