feat(clip): add script to put data into the clipboard
Add a script that takes data and tries to populate the clipboard with
it. If an argument is given, it is assumed to be the file to put in the
X11 clipboard. Otherwise, the contents of stdin will be read to
determine whether they point to a file to be put in the clipboard or are
what should be put there instead.