feat(fzfmenu)!: set override_redirect through patched st
Set the override_redirect bit, center the window, draw a border and
color the border through the flags available in my patched version of
st.

BREAKING CHANGE: with this commit, fzfmenu works only with a patched
version of st that includes the '-b', '-r', '-C' and '-B' flags. Refer
to d3adb5/st on GitHub for the patches.