feat(rop): add run-on-ping script
Add a script that continuously runs a given command if connections are
being successfully closed on a given port. socat is used for listening
for connections, and tools like netcat and anything else, really, can be
used to close the connection.

This is being used as an alternative to a named pipe when running build
and test commands on a different window while changing code with Neovim.
As an example of usage, the makeprg variable in Neovim is set to "nc -z
localhost 8989".