feat(volume): rewrite volume script using notify-send
Rewrite volume script to separate function responsibilities and use
notify-send instead of dunstify, as it is possible to replace
notifications using it.

Additionally, hints are now passed to the notification daemon to pick a
GTK image for the notification, mark the notification as transient, and
indicate a percentage which is normally interpreted as a progress bar.